refactor(handler): split MCP login routes into dedicated handler
